+++ src/main/java/com/arge/issuemanagement/service/impl/ProjectServiceImpl.java
@@ -1,15 +1,21 @@
 package com.arge.issuemanagement.service.impl;
 
+import com.arge.issuemanagement.dto.IssueDto;
 import com.arge.issuemanagement.dto.ProjectDto;
+import com.arge.issuemanagement.entity.Issue;
 import com.arge.issuemanagement.entity.Project;
 import com.arge.issuemanagement.repository.ProjectRepository;
 import com.arge.issuemanagement.service.ProjectService;
+import com.arge.issuemanagement.util.TPage;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Created by dilara.akbas on Sep, 2020
@@ -39,11 +45,19 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     public ProjectDto getById(Long id) {
+
+        if (projectRepository.findById(id).equals(Optional.empty()))
+            throw new IllegalArgumentException("Project Does Not Exist with ID: " + id);
+
         return modelMapper.map(projectRepository.getOne(id), ProjectDto.class);
     }
 
     @Override
     public ProjectDto getByProjectCode(String projectCode) {
+
+        if (projectRepository.findByProjectCode(projectCode).equals(Optional.empty()))
+            throw new IllegalArgumentException("Project Does Not Exist with Project Code: " + projectCode);
+
         return modelMapper.map(projectRepository.getByProjectCode(projectCode), ProjectDto.class);
     }
 
@@ -53,8 +67,12 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
-    public Page<ProjectDto> getAllPageable(Pageable pageable) {
-        return null;
+    public TPage<ProjectDto> getAllPageable(Pageable pageable) {
+
+        Page<Project> data = projectRepository.findAll(pageable);
+        TPage<ProjectDto> response = new TPage<>();
+        response.setStat(data, Arrays.asList((modelMapper.map(data.getContent(), ProjectDto[].class))));
+        return response;
     }
 
     @Override
@@ -63,6 +81,10 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     public Boolean delete(Long id) {
+
+        if (projectRepository.findById(id).equals(Optional.empty()))
+            throw new IllegalArgumentException("Project Does Not Exist with ID: " + id);
+
         projectRepository.deleteById(id);
         return true;
     }
@@ -73,7 +95,7 @@ public class ProjectServiceImpl implements ProjectService {
         Project project = projectRepository.getOne(id);
 
         if (project == null)
-            throw new IllegalArgumentException("Project does not exist with Id: " + id);
+            throw new IllegalArgumentException("Project Does Not Exist with Id: " + id);
 
         Project projectCheck = projectRepository.getByProjectCodeAndIdNot(projectDto.getProjectCode(), id);
